requirements.txt
# AIWindows Client - Python Requirements
# Install via: pip install -r requirements.txt

# GUI Framework
PyQt6>=6.4.0
PyQt6-WebEngine>=6.4.0

# System monitoring
psutil>=5.9.0

# HTTP/API client
requests>=2.28.0
httpx>=0.24.0
aiohttp>=3.8.0

# WebSocket for MCP Node
websockets>=11.0

# Data validation
pydantic>=2.0.0

# Terminal emulation
pyte>=0.8.0
pywinpty>=2.0.0  # Windows ConPTY

# Encryption for secure settings storage
cryptography>=41.0.0

# Async support
anyio>=4.0.0

# Image processing (for icons)
pillow>=10.0.0
